feat(core): change Integer quantity fields to BigDecimal across all domains
- Convert all Integer quantity fields to BigDecimal (decimal(19,3)) in:
  Inventory, Order, Shipping, Receiving, Requisition, Product, Picklist domains
- Update Liquibase migration for all affected tables
- Fix service layer casts (.toInteger() → .toBigDecimal(), as Integer → as BigDecimal)
- Fix controller parameter type casts
- Update method return types for getQuantity*() methods

Issue: #1

+++ grails-app/controllers/org/pih/warehouse/order/OrderController.groovy
@@ -769,7 +769,7 @@ class OrderController {
             if (!orderService.isOrderEditable(orderItem.order, session.user)) {
                 throw new UnsupportedOperationException("${warehouse.message(code: 'errors.noPermissions.label')}")
             }
-            if (params.quantity && orderItem.quantity != (params.quantity as Integer)) {
+            if (params.quantity && orderItem.quantity != (params.quantity as BigDecimal)) {
                 // if existing item's quantity is edited we have to trigger the order summary refresh
                 orderItem.disableRefresh = false
             }
